Moved existing detail pickers to new folder.
