+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/create-post.js
@@ -8,9 +8,11 @@ import "ng-redux";
 import labelledHrModule from "./labelled-hr.js";
 import imageDropzoneModule from "./image-dropzone.js";
 import descriptionPickerModule from "./description-picker.js";
-import locationPickerModule from "./location-picker.js";
-import tagsPickerModule from "./tags-picker.js";
-import ttlPickerModule from "./ttl-picker.js";
+//import descriptionPickerModule from "./details/description-picker.js";
+import locationPickerModule from "./details/location-picker.js";
+//import routePickerModule from "./details/route-picker.js";
+import tagsPickerModule from "./details/tags-picker.js";
+import ttlPickerModule from "./details/ttl-picker.js";
 import createIsseeksModule from "./create-isseeks.js";
 import { postTitleCharacterLimit } from "config";
 import { getIn, attach } from "../utils.js";
