+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/details/location-picker.js
