+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/details/tags-picker.js
